B-23451:moved appendTimestampToFilename to src/utils dir and cleaneup the code.
