+++ src/pages/MyMove/PPM/Closeout/Expenses/Expenses.jsx
@@ -22,7 +22,8 @@ import {
 } from 'services/internalApi';
 import { updateMTOShipment } from 'store/entities/actions';
 import { formatDateForSwagger } from 'shared/dates';
-import { convertDollarsToCents, appendTimestampToFilename } from 'shared/utils';
+import { convertDollarsToCents } from 'shared/utils';
+import appendTimestampToFilename from 'utils/fileUpload';
 
 const Expenses = () => {
   const [errorMessage, setErrorMessage] = useState(null);
